🧪 improve snap-mem.py tests & fix lint CI
- Consolidate snap-mem tests into Cachyos/Scripts/WIP/test_snap_mem.py
- Add test cases for build_base_name (whitespace, invalid dates)
- Add test cases for make_unique_name (collisions, empty sets)
- Fix CI failure by implementing missing lint-format.sh script
- Apply project-wide formatting via lint-format.sh
