+++ .github/workflows/jules-weekly-cleanup.yml
@@ -3,7 +3,7 @@ name: Weekly Codebase Cleanup
 on:
   schedule:
     # Run every Monday at 2 AM UTC
-    - cron: '0 2 * * 1'
+    - cron: "0 2 * * 1"
   workflow_dispatch: # Allow manual triggers
 
 jobs:
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     permissions:
       contents: read
-    
+
     steps:
       - name: Invoke Jules for Cleanup
         uses: BeksOmega/jules-action@v1.0.0
