+++ Cachyos/Scripts/Android/BACKLOG.md
@@ -4,26 +4,30 @@ Tracked feature requests and future work items.
 
 ## Action Items
 
-| Priority | Description | Complexity |
-|----------|-------------|------------|
-| Medium | Fork and optimize [charwasp/modify](https://github.com/charwasp/modify/) — merge into a single Python script<br>and add GraphicsMagick support | Complex |
+| Priority | Description                                                                                                                                    | Complexity |
+| -------- | ---------------------------------------------------------------------------------------------------------------------------------------------- | ---------- |
+| Medium   | Fork and optimize [charwasp/modify](https://github.com/charwasp/modify/) — merge into a single Python script<br>and add GraphicsMagick support | Complex    |
 
 ## Resources to Explore
 
 **Modules & Frameworks:**
+
 - [LSPosed Modules Repository](https://modules.lsposed.org/)
 - [VirtualXposed](https://github.com/android-hacker/VirtualXposed)
 
 **UI/Navigation Fixes:**
+
 - [Fullscreen immersive gesture script](https://github.com/trinhnv1205/optimize/blob/main/menu/fullscreen_immersive_getsure.sh) — may fix 3-button navigation bar
 
 **Modified Apps:**
+
 - [SpotifyPlus](https://github.com/LeNerd46/SpotifyPlus)
 - [NPatch](https://github.com/7723mod/NPatch)
 - [InstaEclipse](https://github.com/ReSo7200/InstaEclipse)
 - [NagramX](https://github.com/risin42/NagramX)
 
 **Development Tools:**
+
 - [AntiSplit-M](https://github.com/pumPCin/AntiSplit-M)
 - [ADB Toolbox](https://github.com/Nahvine/ADB_Toolbox)
 - [ApkPatcher by Th3-R3p4ck3r](https://github.com/Th3-R3p4ck3r/ApkPatcher)
