+++ Cachyos/Scripts/WIP/test_snap_mem.py
@@ -1,6 +1,7 @@
 import unittest
 import importlib.util
 import sys
+import threading
 from pathlib import Path
 
 # Import snap-mem.py using importlib because of the hyphen in the filename
@@ -36,6 +37,48 @@ class TestSnapMem(unittest.TestCase):
     def test_build_base_name_empty(self):
         with self.assertRaises(ValueError):
             snap_mem.build_base_name("")
+
+    def test_build_base_name_whitespace(self):
+        # build_base_name uses datetime.strptime(date_str, DATE_FMT)
+        # strptime is strict about whitespace unless included in fmt
+        date_str = " 2023-01-01 12:00:00 UTC "
+        with self.assertRaises(ValueError):
+            snap_mem.build_base_name(date_str)
+
+    def test_build_base_name_logical_invalid_date(self):
+        # 2023 is not a leap year
+        date_str = "2023-02-29 12:00:00 UTC"
+        with self.assertRaises(ValueError):
+            snap_mem.build_base_name(date_str)
+
+    def test_make_unique_name(self):
+        # make_unique_name updates the 'existing' set in-place
+        existing = {"base.jpg", "base_1.jpg"}
+        lock = threading.Lock()
+
+        # base.jpg is in existing. n=1 -> base_1.jpg (in existing). n=2 -> base_2.jpg.
+        name = snap_mem.make_unique_name("base", ".jpg", existing, lock)
+        self.assertEqual(name, "base_2.jpg")
+        self.assertIn("base_2.jpg", existing)
+
+        # base_3.jpg
+        name2 = snap_mem.make_unique_name("base", ".jpg", existing, lock)
+        self.assertEqual(name2, "base_3.jpg")
+        self.assertIn("base_3.jpg", existing)
+
+    def test_make_unique_name_empty_set(self):
+        existing = set()
+        lock = threading.Lock()
+        name = snap_mem.make_unique_name("base", ".jpg", existing, lock)
+        self.assertEqual(name, "base.jpg")
+        self.assertIn("base.jpg", existing)
+
+    def test_make_unique_name_multiple_collisions(self):
+        existing = {"base.jpg", "base_1.jpg", "base_2.jpg", "base_3.jpg"}
+        lock = threading.Lock()
+        name = snap_mem.make_unique_name("base", ".jpg", existing, lock)
+        self.assertEqual(name, "base_4.jpg")
+        self.assertIn("base_4.jpg", existing)
 
 if __name__ == '__main__':
     unittest.main()
