+++ docs/PERFORMANCE.md
@@ -131,14 +131,14 @@ fi
 
 ### Command Costs (Relative)
 
-| Operation | Cost | Alternative | Cost |
-|-----------|------|-------------|------|
-| `$(command)` | 100x | `${var//pattern/}` | 1x |
-| `tr` | 50x | `${var,,}` | 1x |
-| `basename` | 30x | `${var##*/}` | 1x |
-| `dirname` | 30x | `${var%/*}` | 1x |
-| `cat file` | 20x | `$(<file)` | 1x |
-| `sudo sh -c` | 80x | `printf \| sudo tee` | 40x |
+| Operation    | Cost | Alternative          | Cost |
+| ------------ | ---- | -------------------- | ---- |
+| `$(command)` | 100x | `${var//pattern/}`   | 1x   |
+| `tr`         | 50x  | `${var,,}`           | 1x   |
+| `basename`   | 30x  | `${var##*/}`         | 1x   |
+| `dirname`    | 30x  | `${var%/*}`          | 1x   |
+| `cat file`   | 20x  | `$(<file)`           | 1x   |
+| `sudo sh -c` | 80x  | `printf \| sudo tee` | 40x  |
 
 ## Linting Rules
